requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
Werkzeug==3.0.1